server/src/main/chat/requirements.txt
httpx
fastapi
pydantic
uvicorn
python-dotenv
fastapi_security
motor
asyncio
aio_pika
pika